move main window code
