+++ app/renderer/main/app.html
@@ -3,7 +3,6 @@
   <head>
     <meta charset="utf-8">
     <title>Timesheets</title>
-    <!-- <link rel="stylesheet" href="https://code.getmdl.io/1.1.3/material.cyan-light_blue.min.css" type="text/css"> -->
     <script>
       (function() {
         if (!process.env.HOT) {
