+++ app/renderer/main/components/App/index.jsx
