+++ app/renderer/main/components/Home/index.jsx
