+++ app/renderer/main/containers/App.js
@@ -1,7 +1,7 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import App from '../components/App';
-import * as JobActions from '../../shared/actions/job';
+import * as JobActions from '../../../shared/actions/job';
 
 function mapStateToProps({ job }) {
   return { job };
