+++ app/renderer/main/containers/ProjectsPage.js
@@ -1,8 +1,8 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import Projects from '../components/Projects';
-import * as ProjectActions from '../../shared/actions/project';
-import * as JobActions from '../../shared/actions/job';
+import * as ProjectActions from '../../../shared/actions/project';
+import * as JobActions from '../../../shared/actions/job';
 import { push } from 'react-router-redux';
 
 function mapStateToProps({ project }) {
