+++ app/shared/store/configureStore.development.js
@@ -9,7 +9,7 @@ import getRootReducer from '../reducers';
 import forwardToMain from './middleware/forwardToMain';
 import forwardToRenderer from './middleware/forwardToRenderer';
 import triggerAlias from './middleware/triggerAlias';
-import DevTools from '../../renderer/components/DevTools';
+import DevTools from '../../renderer/main/components/DevTools';
 
 export default function configureStore(initialState, scope = 'main') {
   const logger = createLogger({
