+++ webpack.config.development.js
@@ -11,7 +11,7 @@ const config = {
 
   entry: [
     'webpack-hot-middleware/client?path=http://localhost:3000/__webpack_hmr',
-    './app/renderer/index',
+    './app/renderer/main/index',
   ],
 
   output: {
